pipeline/requirements-dev.txt
boto3==1.10.45
pytest==5.3.2
pyyaml==5.3